refine base pipeline
