+++ cocos/core/game.ts
@@ -634,7 +634,7 @@ export class Game extends EventTarget {
                     console.warn(err);
                 }
                 else {
-                    this.setRenderPipeline(asset);
+                    //this.setRenderPipeline(asset);
                 }
                 this._safeEmit(Game.EVENT_GAME_INITED);
                 if (useSplash) {
