+++ cocos/core/pipeline/forward/scene-culling.ts
@@ -1,0 +1,161 @@
+import { intersect, sphere } from '../../geometry';
+import { Model, Camera } from '../../renderer';
+import { Layers } from '../../scene-graph';
+import { DirectionalLight } from '../../renderer/scene/directional-light';
+import { cullDirectionalLight, cullSphereLight, cullSpotLight } from '../culling';
+import { Vec3} from '../../math';
+import { LightType } from '../../renderer/scene/light';
+import { SKYBOX_FLAG } from '../../renderer';
+import { SphereLight } from '../../renderer/scene/sphere-light';
+import { SpotLight } from '../../renderer/scene/spot-light';
+import { UBOForwardLight } from '../define';
+import { legacyCC } from '../../global-exports';
+import { ForwardRenderContext } from './forward-render-context';
+import { RenderView } from '../render-view';
+import { GFXBufferUsageBit, GFXMemoryUsageBit } from '../../gfx/define';
+import { GFXBuffer } from '../../gfx/buffer';
+
+const _tempVec3 = new Vec3();
+const _sphere = sphere.create(0, 0, 0, 1);
+
+function addVisibleModel (rctx: ForwardRenderContext, model: Model, camera: Camera) {
+    let depth = 0;
+    if (model.node) {
+        Vec3.subtract(_tempVec3, model.node.worldPosition, camera.position);
+        depth = Vec3.dot(_tempVec3, camera.forward);
+    }
+    rctx.renderObjects.push({
+        model,
+        depth,
+    });
+}
+
+function cullLightPerModel (rctx: ForwardRenderContext, model: Model) {
+    const validLights = rctx.validLights;
+    const lightIndices = rctx.lightIndices;
+
+    if (model.node) {
+        model.node.getWorldPosition(_tempVec3);
+    } else {
+        _tempVec3.set(0.0, 0.0, 0.0);
+    }
+    for (let i = 0; i < validLights.length; i++) {
+        let isCulled = false;
+        switch (validLights[i].type) {
+            case LightType.DIRECTIONAL:
+                isCulled = cullDirectionalLight(validLights[i] as DirectionalLight, model);
+                break;
+            case LightType.SPHERE:
+                isCulled = cullSphereLight(validLights[i] as SphereLight, model);
+                break;
+            case LightType.SPOT:
+                isCulled = cullSpotLight(validLights[i] as SpotLight, model);
+                break;
+        }
+        if (!isCulled) {
+            lightIndices.push(i);
+        }
+    }
+}
+
+export function sceneCulling (rctx: ForwardRenderContext, view: RenderView) {
+    const camera = view.camera;
+    const scene = camera.scene!;
+    const device = rctx.device;
+    const renderObjects = rctx.renderObjects;
+    renderObjects.length = 0;
+
+    const mainLight = scene.mainLight;
+    const planarShadows = scene.planarShadows;
+    if (mainLight) {
+        mainLight.update();
+        if (planarShadows.enabled && mainLight.node!.hasChangedFlags) {
+            planarShadows.updateDirLight(mainLight);
+        }
+    }
+
+    if (scene.skybox.enabled && (camera.clearFlag & SKYBOX_FLAG)) {
+        addVisibleModel(rctx, scene.skybox, camera);
+    }
+
+    const models = scene.models;
+    const stamp = legacyCC.director.getTotalFrames();
+
+    for (let i = 0; i < models.length; i++) {
+        const model = models[i];
+
+        // filter model by view visibility
+        if (model.enabled) {
+            const vis = view.visibility & Layers.BitMask.UI_2D;
+            if (vis) {
+                if ((model.node && (view.visibility === model.node.layer)) ||
+                    view.visibility === model.visFlags) {
+                    model.updateTransform(stamp);
+                    model.updateUBOs(stamp);
+                    addVisibleModel(rctx, model, camera);
+                }
+            } else {
+                if (model.node && ((view.visibility & model.node.layer) === model.node.layer) ||
+                    (view.visibility & model.visFlags)) {
+                    model.updateTransform(stamp);
+
+                    // frustum culling
+                    if (model.worldBounds && !intersect.aabb_frustum(model.worldBounds, camera.frustum)) {
+                        continue;
+                    }
+
+                    model.updateUBOs(stamp);
+                    addVisibleModel(rctx, model, camera);
+                }
+            }
+        }
+    }
+
+    if (planarShadows.enabled) {
+        planarShadows.updateShadowList(camera.frustum, stamp, (camera.visibility & Layers.BitMask.DEFAULT) !== 0);
+    }
+
+    const validLights = rctx.validLights;
+    const lightBuffers = rctx.lightBuffers;
+    const lightIndexOffsets = rctx.lightIndexOffsets;
+    const lightIndices = rctx.lightIndices;
+    validLights.length = lightIndexOffsets.length = lightIndices.length = 0;
+    const sphereLights = view.camera.scene!.sphereLights;
+
+    for (let i = 0; i < sphereLights.length; i++) {
+        const light = sphereLights[i];
+        light.update();
+        sphere.set(_sphere, light.position.x, light.position.y, light.position.z, light.range);
+        if (intersect.sphere_frustum(_sphere, view.camera.frustum)) {
+            validLights.push(light);
+        }
+    }
+    const spotLights = view.camera.scene!.spotLights;
+    for (let i = 0; i < spotLights.length; i++) {
+        const light = spotLights[i];
+        light.update();
+        sphere.set(_sphere, light.position.x, light.position.y, light.position.z, light.range);
+        if (intersect.sphere_frustum(_sphere, view.camera.frustum)) {
+            validLights.push(light);
+        }
+    }
+
+    if (validLights.length > lightBuffers.length) {
+        for (let l = lightBuffers.length; l < validLights.length; ++l) {
+            const lightBuffer: GFXBuffer = device.createBuffer({
+                usage: GFXBufferUsageBit.UNIFORM | GFXBufferUsageBit.TRANSFER_DST,
+                memUsage: GFXMemoryUsageBit.HOST | GFXMemoryUsageBit.DEVICE,
+                size: UBOForwardLight.SIZE,
+                stride: UBOForwardLight.SIZE,
+            });
+            lightBuffers.push(lightBuffer);
+        }
+    }
+
+    if (validLights.length > 0) {
+        for (let i = 0; i < renderObjects.length; i++) {
+            lightIndexOffsets[i] = lightIndices.length;
+            cullLightPerModel(rctx, renderObjects[i].model);
+        }
+    }
+}
