+++ cocos/core/root.ts
@@ -287,10 +287,6 @@ export class Root {
             }
         }
 
-        if (this._pipeline) {
-            this._pipeline.resize(width, height);
-        }
-
         for (const view of this._views) {
             if (view.camera.isWindowSize) {
                 view.camera.resize(width, height);
